AD总有问题，原来是Clock应该*64就好了，可以查看电压以及输出串口了

+++ Quadrotor/system/src/quadro_app/variables.c
@@ -1,6 +1,6 @@
 #include "variables.h"
 
-long AdValueGyroNick, AdValueGyroRoll, AdValueGyroYaw,AdValueAccRoll, AdValueAccNick , AdValueAccTop;
+float AdValueGyroNick, AdValueGyroRoll, AdValueGyroYaw,AdValueAccRoll, AdValueAccNick , AdValueAccTop;
 volatile unsigned char ADReady=0,ADTFinish=1;
 unsigned char msg[COMMAND_MAX_LENGTH_IN_BYTES];
 volatile unsigned short voltage[16] = { 0 };
