+++ Quadrotor/system/src/quadro_app/variables.h
@@ -1,5 +1,5 @@
 #define COMMAND_MAX_LENGTH_IN_BYTES 14
-extern long AdValueGyroNick, AdValueGyroRoll, AdValueGyroYaw,AdValueAccRoll, AdValueAccNick , AdValueAccTop;
+extern float AdValueGyroNick, AdValueGyroRoll, AdValueGyroYaw,AdValueAccRoll, AdValueAccNick , AdValueAccTop;
 extern volatile unsigned char ADReady,ADTFinish;
 extern unsigned char msg[COMMAND_MAX_LENGTH_IN_BYTES];
 extern volatile unsigned short voltage[16];
